# docs/swagger.yaml
basePath: /api
definitions:
...
  GAMERS-BE_internal_contest_application_dto.CreateContestRequest:
    properties:
      auto_start:
        type: boolean
      contest_type:
        $ref: '#/definitions/GAMERS-BE_internal_contest_domain.ContestType'
      description:
        type: string
      discord_guild_id:
        type: string
      discord_text_channel_id:
        type: string
      ended_at:
        type: string
      game_point_table_id:
        type: integer
      game_type:
        $ref: '#/definitions/GAMERS-BE_internal_game_domain.GameType'
      max_team_count:
        type: integer
      started_at:
        type: string
      thumbnail:
        type: string
      title:
        type: string
      total_point:
        type: integer
      total_team_member:
        type: integer
    required:
    - contest_type
    - title
    type: object
  GAMERS-BE_internal_contest_application_dto.DiscordLinkRequiredResponse:
    properties:
      message:
        type: string
      oauth_url:
        type: string
    type: object
...
  GAMERS-BE_internal_game_domain.GameType:
    enum:
    - VALORANT
    - LOL
    type: string
    x-enum-varnames:
    - GameTypeValorant
    - GameTypeLOL
  GAMERS-BE_internal_global_common_dto.PaginationResponse:
    properties:
      data: {}
      page:
        type: integer
      page_size:
        type: integer
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
  GAMERS-BE_internal_global_response.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@gamers.com
    name: API Support
  description: API Server for GAMERS platform
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: GAMERS API
  version: "1.0"
paths:
...
  /api/contests:
    get:
      consumes:
      - application/json
      description: Get all contests with pagination, sorting, and title search support
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 10, max: 100)'
        in: query
        name: page_size
        type: integer
      - description: 'Sort field: created_at, started_at, ended_at (default: created_at)'
        in: query
        name: sort_by
        type: string
      - description: 'Sort order: asc, desc (default: desc)'
        in: query
        name: order
        type: string
      - description: Search by contest title (partial match)
        in: query
        name: title
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/GAMERS-BE_internal_global_common_dto.PaginationResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
      summary: Get all contests with pagination
      tags:
      - contests
    post:
      consumes:
      - application/json
      description: Create a new contest with contest details
      parameters:
      - description: Contest creation request
        in: body
        name: contest
        required: true
        schema:
          $ref: '#/definitions/GAMERS-BE_internal_contest_application_dto.CreateContestRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/GAMERS-BE_internal_contest_application_dto.ContestResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/GAMERS-BE_internal_contest_application_dto.DiscordLinkRequiredResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Create a new contest
      tags:
      - contests
...
  /api/contests/{contestId}/games/{gameId}/detect:
    post:
      description: Staff manually triggers match detection for a specific game
      parameters:
      - description: Contest ID
        in: path
        name: contestId
        required: true
        type: integer
      - description: Game ID
        in: path
        name: gameId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
      security:
      - BearerAuth: []
      summary: Manually trigger match detection
      tags:
      - games
      - match-detection
  /api/contests/{contestId}/games/{gameId}/detection-status:
    get:
      description: Returns the current detection status for a game
      parameters:
      - description: Contest ID
        in: path
        name: contestId
        required: true
        type: integer
      - description: Game ID
        in: path
        name: gameId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/GAMERS-BE_internal_game_application_dto.ScheduleGameResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
      summary: Get match detection status
      tags:
      - games
      - match-detection
  /api/contests/{contestId}/games/{gameId}/result:
    get:
      description: Returns the match result for a finished game
      parameters:
      - description: Contest ID
        in: path
        name: contestId
        required: true
        type: integer
      - description: Game ID
        in: path
        name: gameId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/GAMERS-BE_internal_game_application_dto.MatchResultResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
      summary: Get match result
      tags:
      - games
      - match-detection
    post:
      consumes:
      - application/json
      description: Staff manually inputs the game result (fallback when auto-detection
        fails)
      parameters:
      - description: Contest ID
        in: path
        name: contestId
        required: true
        type: integer
      - description: Game ID
        in: path
        name: gameId
        required: true
        type: integer
      - description: Manual result request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/GAMERS-BE_internal_game_application_dto.ManualResultRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/GAMERS-BE_internal_game_application_dto.MatchResultResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
      security:
      - BearerAuth: []
      summary: Submit manual game result
      tags:
      - games
      - match-detection
  /api/contests/{contestId}/games/{gameId}/result/stats:
    get:
      description: Returns the match result and individual player stats
      parameters:
      - description: Contest ID
        in: path
        name: contestId
        required: true
        type: integer
      - description: Game ID
        in: path
        name: gameId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/GAMERS-BE_internal_game_application_dto.MatchResultResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
      summary: Get match result with player stats
      tags:
      - games
      - match-detection
  /api/contests/{contestId}/games/{gameId}/schedule:
    put:
      consumes:
      - application/json
      description: Staff sets the scheduled start time and detection window for a
        tournament game
      parameters:
      - description: Contest ID
        in: path
        name: contestId
        required: true
        type: integer
      - description: Game ID
        in: path
        name: gameId
        required: true
        type: integer
      - description: Schedule request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/GAMERS-BE_internal_game_application_dto.ScheduleGameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/GAMERS-BE_internal_game_application_dto.ScheduleGameResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
      security:
      - BearerAuth: []
      summary: Set game scheduled start time
      tags:
      - games
      - match-detection
...
  /api/contests/{id}:
    delete:
      consumes:
      - application/json
      description: Delete contest by contest ID
      parameters:
      - description: Contest ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
      security:
      - BearerAuth: []
      summary: Delete a contest
      tags:
      - contests
    get:
      consumes:
      - application/json
      description: Get contest details by contest ID
      parameters:
      - description: Contest ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/GAMERS-BE_internal_contest_application_dto.ContestResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
      summary: Get a contest by ID
      tags:
      - contests
    patch:
      consumes:
      - application/json
      description: Update contest details by contest ID
      parameters:
      - description: Contest ID
        in: path
        name: id
        required: true
        type: integer
      - description: Contest update request
        in: body
        name: contest
        required: true
        schema:
          $ref: '#/definitions/GAMERS-BE_internal_contest_application_dto.UpdateContestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/GAMERS-BE_internal_contest_application_dto.ContestResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
      security:
      - BearerAuth: []
      summary: Update a contest
      tags:
      - contests
...
  /api/contests/{id}/games:
    get:
      consumes:
      - application/json
      description: Get all games under a specific contest
      parameters:
      - description: Contest ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/GAMERS-BE_internal_game_application_dto.GameResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
      summary: Get all games for a contest
      tags:
      - games
  /api/contests/{id}/result:
    get:
      description: Returns the full tournament bracket with game results and champion
      parameters:
      - description: Contest ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/GAMERS-BE_internal_game_application_dto.ContestResultResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
      summary: Get tournament contest result
      tags:
      - games
      - tournaments
...
  /api/games:
    post:
      consumes:
      - application/json
      description: Create a new game under a contest (creator becomes the leader)
      parameters:
      - description: Game creation request
        in: body
        name: game
        required: true
        schema:
          $ref: '#/definitions/GAMERS-BE_internal_game_application_dto.CreateGameRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/GAMERS-BE_internal_game_application_dto.GameResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
      security:
      - BearerAuth: []
      summary: Create a new game
      tags:
      - games
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"